// dto/order.dto.ts
export interface OrderDto {
  order: Order;
}

interface Order {
  id: number;
  admin_graphql_api_id: string;
  app_id?: any;
  browser_ip: string;
  buyer_accepts_marketing: boolean;
  cancel_reason?: any;
  cancelled_at?: any;
  cart_token: string;
  checkout_id: number;
  checkout_token: string;
  client_details: Clientdetails;
  closed_at?: any;
  confirmed: boolean;
  contact_email: string;
  created_at: string;
  currency: string;
  current_subtotal_price: string;
  current_subtotal_price_set: Currentsubtotalpriceset;
  current_total_discounts: string;
  current_total_discounts_set: Currentsubtotalpriceset;
  current_total_duties_set?: any;
  current_total_price: string;
  current_total_price_set: Currentsubtotalpriceset;
  current_total_tax: string;
  current_total_tax_set: Currentsubtotalpriceset;
  customer_locale?: any;
  device_id?: any;
  discount_codes: Discountcode[];
  email: string;
  estimated_taxes: boolean;
  financial_status: string;
  fulfillment_status?: any;
  gateway: string;
  landing_site: string;
  landing_site_ref: string;
  location_id?: any;
  name: string;
  note?: any;
  note_attributes: Noteattribute[];
  number: number;
  order_number: number;
  order_status_url: string;
  original_total_duties_set?: any;
  payment_gateway_names: string[];
  phone: string;
  presentment_currency: string;
  processed_at: string;
  processing_method: string;
  reference: string;
  referring_site: string;
  source_identifier: string;
  source_name: string;
  source_url?: any;
  subtotal_price: string;
  subtotal_price_set: Currentsubtotalpriceset;
  tags: string;
  tax_lines: Taxline[];
  taxes_included: boolean;
  test: boolean;
  token: string;
  total_discounts: string;
  total_discounts_set: Currentsubtotalpriceset;
  total_line_items_price: string;
  total_line_items_price_set: Currentsubtotalpriceset;
  total_outstanding: string;
  total_price: string;
  total_price_set: Currentsubtotalpriceset;
  total_price_usd: string;
  total_shipping_price_set: Currentsubtotalpriceset;
  total_tax: string;
  total_tax_set: Currentsubtotalpriceset;
  total_tip_received: string;
  total_weight: number;
  updated_at: string;
  user_id?: any;
  billing_address: Billingaddress;
  customer: Customer;
  discount_applications: Discountapplication[];
  fulfillments: Fulfillment[];
  line_items: Lineitem2[];
  payment_details: Paymentdetails;
  refunds: Refund[];
  shipping_address: Billingaddress;
  shipping_lines: Shippingline[];
}

interface Shippingline {
  id: number;
  carrier_identifier?: any;
  code: string;
  delivery_category?: any;
  discounted_price: string;
  discounted_price_set: Currentsubtotalpriceset;
  phone?: any;
  price: string;
  price_set: Currentsubtotalpriceset;
  requested_fulfillment_service_id?: any;
  source: string;
  title: string;
  tax_lines: any[];
  discount_allocations: any[];
}

interface Refund {
  id: number;
  admin_graphql_api_id: string;
  created_at: string;
  note: string;
  order_id: number;
  processed_at: string;
  restock: boolean;
  total_additional_fees_set: Currentsubtotalpriceset;
  total_duties_set: Currentsubtotalpriceset;
  user_id: number;
  order_adjustments: any[];
  transactions: Transaction[];
  refund_line_items: Refundlineitem[];
  duties: any[];
  additional_fees: any[];
}
// ...
interface Paymentdetails {
  credit_card_bin?: any;
  avs_result_code?: any;
  cvv_result_code?: any;
  credit_card_number: string;
  credit_card_company: string;
  credit_card_name?: any;
  credit_card_wallet?: any;
  credit_card_expiration_month?: any;
  credit_card_expiration_year?: any;
}

interface Lineitem2 {
  id: number;
  admin_graphql_api_id: string;
  fulfillable_quantity: number;
  fulfillment_service: string;
  fulfillment_status?: any;
  gift_card: boolean;
  grams: number;
  name: string;
  price: string;
  price_set: Currentsubtotalpriceset;
  product_exists: boolean;
  product_id: number;
  properties: Noteattribute[];
  quantity: number;
  requires_shipping: boolean;
  sku: string;
  taxable: boolean;
  title: string;
  total_discount: string;
  total_discount_set: Currentsubtotalpriceset;
  variant_id: number;
  variant_inventory_management: string;
  variant_title: string;
  vendor?: any;
  tax_lines: Taxline[];
  duties: any[];
  discount_allocations: Discountallocation[];
}

interface Fulfillment {
  id: number;
  admin_graphql_api_id: string;
  created_at: string;
  location_id: number;
  name: string;
  order_id: number;
  origin_address: Originaddress;
  receipt: Receipt;
  service: string;
  shipment_status?: any;
  status: string;
  tracking_company: string;
  tracking_number: string;
  tracking_numbers: string[];
  tracking_url: string;
  tracking_urls: string[];
  updated_at: string;
  line_items: Lineitem[];
}
// ...
interface Discountapplication {
  target_type: string;
  type: string;
  value: string;
  value_type: string;
  allocation_method: string;
  target_selection: string;
  code: string;
}

interface Customer {
  id: number;
  email: string;
  accepts_marketing: boolean;
  created_at: string;
  updated_at: string;
  first_name: string;
  last_name: string;
  state: string;
  note?: any;
  verified_email: boolean;
  multipass_identifier?: any;
  tax_exempt: boolean;
  phone: string;
  tags: string;
  currency: string;
  accepts_marketing_updated_at: string;
  marketing_opt_in_level?: any;
  tax_exemptions: any[];
  email_marketing_consent: Emailmarketingconsent;
  sms_marketing_consent?: any;
  sms_transactional_consent?: any;
  admin_graphql_api_id: string;
  default_address: Defaultaddress;
}
// ...
interface Billingaddress {
  first_name: string;
  address1: string;
  phone: string;
  city: string;
  zip: string;
  province: string;
  country: string;
  last_name: string;
  address2: string;
  company?: any;
  latitude: number;
  longitude: number;
  name: string;
  country_code: string;
  province_code: string;
}

interface Taxline {
  price: string;
  rate: number;
  title: string;
  price_set: Currentsubtotalpriceset;
  channel_liable?: any;
}

interface Noteattribute {
  name: string;
  value: string;
}

interface Discountcode {
  code: string;
  amount: string;
  type: string;
}

interface Currentsubtotalpriceset {
  shop_money: Shopmoney;
  presentment_money: Shopmoney;
}

interface Shopmoney {
  amount: string;
  currency_code: string;
}

interface Clientdetails {
  accept_language?: any;
  browser_height?: any;
  browser_ip: string;
  browser_width?: any;
  session_hash?: any;
  user_agent?: any;
}